Feat : Player Damaged 매개변수 변경, 배경 order 설정, OneWayPlatform(아래에서 위로만 가능), 몬스터 데미지 20설정, 초반 몬스터 10으로 설정
