+++ Assets/Prefabs/Player/Player.prefab
@@ -761,6 +761,7 @@ GameObject:
   - component: {fileID: 2612563571394785668}
   - component: {fileID: 7770274658885052061}
   - component: {fileID: 2016939861136809588}
+  - component: {fileID: 8150052083277850353}
   m_Layer: 11
   m_Name: Player
   m_TagString: Player
@@ -778,8 +779,8 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: -1, z: -0.7}
-  m_LocalScale: {x: 1.5, y: 1.5, z: 1}
-  m_ConstrainProportionsScale: 0
+  m_LocalScale: {x: 1.5, y: 1.5, z: 1.5}
+  m_ConstrainProportionsScale: 1
   m_Children:
   - {fileID: 1688892288499030834}
   m_Father: {fileID: 0}
@@ -989,6 +990,18 @@ Animator:
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorStateOnDisable: 0
   m_WriteDefaultValuesOnDisable: 0
+--- !u!114 &8150052083277850353
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7418917979669633661}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 03d9ac30c8d964943882bcd04e4b7bd0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &8064527702958576006
 GameObject:
   m_ObjectHideFlags: 0
