+++ Assets/Scripts/Enemy/EnemySpecific/Archer/Data/Archer_RangeAttackStateData.asset
@@ -13,6 +13,6 @@ MonoBehaviour:
   m_Name: Archer_RangeAttackStateData
   m_EditorClassIdentifier: 
   projectileGameObject: {fileID: 4615923376900856792, guid: b485cdb0d71c1844696846f953b7dc26, type: 3}
-  projectileDamage: 10
+  projectileDamage: 20
   projectileSpeed: 50
   projectiletravelDistance: 10
