+++ Assets/Scripts/Enemy/EnemySpecific/Assassin_Cultist/Data/AssassinCultist_MeleeAttackStateData.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: AssassinCultist_MeleeAttackStateData
   m_EditorClassIdentifier: 
   attackRadius: 1
-  attackDamage: 10
+  attackDamage: 20
   whatIsPlayer:
     serializedVersion: 2
     m_Bits: 2048
