+++ Assets/Scripts/Enemy/EnemySpecific/EvilWizard/Data/EvilWizard_MeleeAttack2StateData.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: EvilWizard_MeleeAttack2StateData
   m_EditorClassIdentifier: 
   attackRadius: 2.3
-  attackDamage: 10
+  attackDamage: 20
   whatIsPlayer:
     serializedVersion: 2
     m_Bits: 2048
