+++ Assets/Scripts/Enemy/EnemySpecific/FireWizard/Data/FireWizard_MeleeAttackStateData.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: FireWizard_MeleeAttackStateData
   m_EditorClassIdentifier: 
   attackRadius: 1.2
-  attackDamage: 10
+  attackDamage: 20
   whatIsPlayer:
     serializedVersion: 2
     m_Bits: 2048
