+++ Assets/Scripts/Enemy/EnemySpecific/Mushroom/Data/Mushroom_MeleeAttackStateData.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: Mushroom_MeleeAttackStateData
   m_EditorClassIdentifier: 
   attackRadius: 0.5
-  attackDamage: 0
+  attackDamage: 10
   whatIsPlayer:
     serializedVersion: 2
     m_Bits: 2048
