+++ Assets/Scripts/Enemy/EnemySpecific/TwistedCultist/Data/TwistedCultist_MeleeAttackStateData.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: TwistedCultist_MeleeAttackStateData
   m_EditorClassIdentifier: 
   attackRadius: 0.5
-  attackDamage: 10
+  attackDamage: 20
   whatIsPlayer:
     serializedVersion: 2
     m_Bits: 2048
