+++ Assets/Scripts/Enemy/States/MeleeAttackState.cs
@@ -47,7 +47,7 @@ public class MeleeAttackState : AttackState
             //Debug.Log("플레이어 공격!");
 
             var player = obj.gameObject.GetComponent<PlayerController>();
-            player.Damaged(1, entity.gameObject.transform.position);
+            player.Damaged(1, entity);
         }
     }
