+++ Assets/Scripts/Map/Parallax.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ca44b0775463f574f8759048d928e7c2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
