+++ Assets/Scripts/Map/Parallax/ParallaxCamera.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 24bf9de13f93820458f06a9bcb80bc13
