+++ Assets/Scripts/Map/Parallax/ParallaxLayer.cs
@@ -1,0 +1,16 @@
+using UnityEngine;
+
+[ExecuteInEditMode]
+public class ParallaxLayer : MonoBehaviour
+{
+    public float parallaxFactor;
+
+    public void Move(float delta)
+    {
+        Vector3 newPos = transform.localPosition;
+        newPos.x -= delta * parallaxFactor;
+
+        transform.localPosition = newPos;
+    }
+
+}
