+++ Assets/Scripts/Map/Parallax/ParallaxLayer.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b8a28f989a9f7da439bcc780933fce79
