--- Assets/Scripts/Map/ParallaxBackground.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c0837c7cb7db42b4e8f8a520f3d668cf
