--- Assets/Scripts/Map/ParallaxCamera.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 24bf9de13f93820458f06a9bcb80bc13
