--- Assets/Scripts/Map/ParallaxLayer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b8a28f989a9f7da439bcc780933fce79
