--- Assets/Scripts/Map/Trap.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a2e1a9451268e1d47bb107d79c28e6da
