+++ Assets/Scripts/Player/OneWayPlatform.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 03d9ac30c8d964943882bcd04e4b7bd0
