+++ Assets/Scripts/Trap.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab2284e5533112e4b88d8ccabe09df11
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
