+++ Assets/Scripts/Trap/Trap.cs
@@ -1,0 +1,20 @@
+using UnityEngine;
+
+public class Trap : MonoBehaviour
+{
+    Collider2D coll;
+
+    private void Awake()
+    {
+        coll = GetComponent<Collider2D>();
+    }
+
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        if(collision.gameObject.CompareTag("Player"))
+        {
+            Debug.Log($"{GetType()} : 충돌");
+        }
+
+    }
+}
